Change ViewModel structure and Act5_Fragments behaviour
ViewModel now uses LiveList, allowing for observing data. That means
Fragments will now update on every change in list of points, not only
when fragment is resumed. This will be necessary for dynamic updates of
route in Act5_Frag2 when user changes order of points.
Furthermore, ViewModel stores now LatLng instead for Markers, in
MutablePair wrappers. This makes sure, that data in ViewModel is
independent from Markers on specific maps, which are now stored in lists
inside of Fragments, as they are bound to LifeCycle of MapViews.

MapViews behaviour inside Fragments is now more polished:
 - UI controls no longer appear when Marker is clicked
 - Camera no longer moves when Marker is clicked
 - Map will now only update added/removed Markers (still room for
 optimization improvements)
 - Changing order in draggable list will now allow live changes in maps
 (optimization definitely needed, also need to fire onRowMoved after drag
 is completed)
 - Map in Act5_Frag2 is now displaying all Markers in most optimal way
 - All gesture controls are now disabled on Map in Act5_Frag2
 - Camera zoom is no longer wrongly restricted by initial zoom of MapViews